Prepare Deploy

+++ pages/index.jsx
@@ -1,12 +1,16 @@
 import React, { useState } from 'react';
 import useSWR from 'swr';
+import dynamic from 'next/dynamic';
 
 import { useUser } from '@hooks/useUser';
-import BlogModal from '@components/blog/blogForm';
-import SmallBlog from '@components/blog/smallBlog';
 import fetcher from '@lib/blog';
 import LandingBigBlog from '@components/blog/landingBlog/bigBlog';
 import LandingSmallBlog from '@components/blog/landingBlog/smallBlog';
+import { jsonStringify } from '@lib/json';
+import idGenerator from '@lib/stringGenerator';
+
+const BlogModal = dynamic(() => import('@components/blog/blogModal'));
+const SmallBlog = dynamic(() => import('@components/blog/smallBlog'));
 
 export default function Index() {
   const [isCreateOpen, setIsCreateOpen] = useState(false);
@@ -20,16 +24,47 @@ export default function Index() {
   const deleteBlog = async (id) => {
     const deleteURL = `${process.env.NEXT_PUBLIC_API_URL}/items/${id}`;
 
-    const result = data.filter((blog) => blog.id !== id);
-    mutate(result, false);
-    await fetch(deleteURL, {
-      method: 'DELETE',
+    // eslint-disable-next-line no-alert
+    if (window.confirm('Are you sure you want to delete it?')) {
+      const result = data.filter((blog) => blog.id !== id);
+      mutate(result, false);
+      await fetch(deleteURL, {
+        method: 'DELETE',
+        headers: {
+          Authorization: token,
+          'Content-Type': 'application/json',
+        },
+      });
+    } else mutate(data);
+  };
+
+  const createBlog = async (body) => {
+    const blog = { ...body, owner_id: user.id };
+    mutate([...data, blog], false);
+    await fetch(`${process.env.NEXT_PUBLIC_API_URL}/items/`, {
+      method: 'POST',
       headers: {
         Authorization: token,
         'Content-Type': 'application/json',
       },
+      body: jsonStringify(body),
     });
-    mutate(result);
+    mutate([...data, body]);
+  };
+
+  const updateBlog = async (body) => {
+    const index = data.findIndex((val) => val.id === body.id);
+    data[index].title = body.title;
+    data[index].description = body.description;
+    mutate(data, false);
+    await fetch(`${process.env.NEXT_PUBLIC_API_URL}/items/${body.id}`, {
+      method: 'PUT',
+      headers: {
+        Authorization: token,
+        'Content-Type': 'application/json',
+      },
+      body: jsonStringify(body),
+    });
   };
 
   const handleUpdateSubmit = (postData) => {
@@ -40,7 +75,7 @@ export default function Index() {
   const renderSmallBlogs = () => {
     const remain = data.slice(3);
     return remain.map((val) => (
-      <div key={val.id}>
+      <div key={val?.id || idGenerator()}>
         <SmallBlog
           title={val.title}
           image={`https://avatar.tobi.sh/${val.title}`}
@@ -68,31 +103,61 @@ export default function Index() {
         />
       </div>
       <div className="w-[600px] flex flex-col justify-between">
-        <LandingSmallBlog
-          title={data[1].title}
-          image={`https://avatar.tobi.sh/${data[1].title}`}
-          description={data[1].description}
-          id={data[1].id}
-          deleteBlog={deleteBlog}
-          user={user}
-          handleUpdateSubmit={handleUpdateSubmit}
-        />
-        <LandingSmallBlog
-          title={data[2].title}
-          image={`https://avatar.tobi.sh/${data[1].title}`}
-          description={data[2].description}
-          id={data[2].id}
-          deleteBlog={deleteBlog}
-          user={user}
-          handleUpdateSubmit={handleUpdateSubmit}
-        />
+        {data[1] ? (
+          <LandingSmallBlog
+            title={data[1].title}
+            image={`https://avatar.tobi.sh/${data[1].title}`}
+            description={data[1].description}
+            id={data[1].id}
+            deleteBlog={deleteBlog}
+            user={user}
+            handleUpdateSubmit={handleUpdateSubmit}
+          />
+        ) : null}
+        {data[2] ? (
+          <LandingSmallBlog
+            title={data[2].title}
+            image={`https://avatar.tobi.sh/${data[2].title}`}
+            description={data[2].description}
+            id={data[2].id}
+            deleteBlog={deleteBlog}
+            user={user}
+            handleUpdateSubmit={handleUpdateSubmit}
+          />
+        ) : null}
       </div>
     </div>
   );
 
+  const renderBlogs = () => {
+    if (!data.length)
+      return (
+        <div className="h-[calc(100vh-178px)] w-full grid place-items-center">
+          <div className="text-center">
+            <h2 className="font-bold text-[24px]">No blogs available</h2>
+            <p className="font-semibold text-[20px]">
+              Click on Create to create a blog
+            </p>
+          </div>
+        </div>
+      );
+
+    if (data.length <= 3) return <div>{renderLandingBlogs()}</div>;
+    return (
+      <>
+        <div>{renderLandingBlogs()}</div>
+        <div className="grid grid-cols-3 justify-items-center gap-x-[110px] gap-y-[4rem] mt-[4rem]">
+          {renderSmallBlogs()}
+        </div>
+      </>
+    );
+  };
+
   if (!data && !error) {
     return <div>Loading...</div>;
   }
+
+  console.log(data);
 
   return (
     <div>
@@ -113,6 +178,7 @@ export default function Index() {
       </div>
       <BlogModal
         buttonText="Create Blog"
+        submitCallback={createBlog}
         isOpen={isCreateOpen}
         setIsOpen={setIsCreateOpen}
       />
@@ -120,15 +186,13 @@ export default function Index() {
         <BlogModal
           buttonText="Update Blog"
           globalData={globalData}
+          submitCallback={updateBlog}
           isOpen={isUpdateOpen}
           setIsOpen={setIsUpdateOpen}
           setGlobalData={setGlobalData}
         />
       )}
-      <div>{renderLandingBlogs()}</div>
-      <div className="grid grid-cols-3 justify-items-center gap-x-[110px] gap-y-[4rem] mt-[4rem]">
-        {renderSmallBlogs()}
-      </div>
+      <div>{renderBlogs()}</div>
     </div>
   );
 }
